server restructure
