--- lib/changelog-as-markdown.js
@@ -1,9 +1,0 @@
-const releaseNotes = require("../client/src/lib/docs/release-notes.json");
-const changelogToMarkdown = require("../client/src/lib/changelog-to-markdown");
-const fs = require("fs");
-
-try {
-  fs.writeFileSync("CHANGELOG.md", changelogToMarkdown(releaseNotes));
-} catch (err) {
-  console.error(err);
-}
