+++ package.json
@@ -6,17 +6,17 @@
   "scripts": {
     "test": "mocha unit-tests/ --recursive --reporter-option maxDiffSize=0",
     "coverage": "npx nyc --check-coverage=true npm run test",
-    "build": "npm i && cd client && npm i && npm run build && cp -r build ../",
+    "build": "cd server && npm i && cd ../client && npm i && npm run build && cp -r build ../server/",
     "client": "npm run build && open http://localhost:3000",
-    "start": "NODE_ENV=production && npm run build && node server.js",
+    "start": "NODE_ENV=production && npm run build && node server/server.js",
     "md": "node lib/docs-as-markdown",
     "changelog": "node lib/changelog-as-markdown",
     "deploy": "./deploy.sh",
     "compatibility-tests": "./compatibility.sh",
     "tf": "node lib/cli-json-to-tf.js",
-    "setup": "npm i mocha -g && npm i && cd client && npm i",
+    "setup": "npm i mocha -g && npm i && cd server && npm i && cd ../client && npm i",
     "front-end-test": "cd client && npm test",
-    "i-all": "npm i && cd client && npm i",
+    "i-all": "cd server npm i && cd ../client && npm i",
     "pretty": "npx prettier --write .",
     "dev-start": "cd client && npm start",
     "pre-commit-build": "cd client && npm run build",
@@ -36,28 +36,18 @@
     "Connor Terech"
   ],
   "license": "ISC",
-  "dependencies": {
-    "axios": "^1.7.4",
-    "blob-stream": "^0.1.3",
-    "body-parser": "^2.0.0",
-    "dotenv": "^17.0.0",
-    "express": "^5.0.0",
-    "js-yaml": "^4.1.0",
-    "json-to-tf": "^0.3.0",
-    "lazy-z": "1.12.1",
-    "next": "^15.5.4",
-    "regex-but-with-words": "^1.5.1",
-    "swagger-ui-express": "^5.0.0",
-    "tar-stream": "^3.1.6"
-  },
   "devDependencies": {
-    "chai": "^6.0.0",
+    "blob-stream": "^0.1.3",
+    "chai": "^4.2.0",
+    "form-data": "^4.0.4",
     "husky": "^9.0.0",
+    "lazy-z": "1.12.1",
     "mocha": "^11.0.0",
     "nodemon": "^3.0.1",
     "nyc": "^17.0.0",
     "prettier": "^3.0.0",
-    "sinon": "^21.0.0"
+    "sinon": "^21.0.0",
+    "tar-stream": "^3.1.6"
   },
   "nyc": {
     "branches": 100,
