+++ server/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "craig-server",
+  "description": "server backend for generating ibm cloud infrastructure resources",
+  "main": "index.js",
+  "scripts": {
+    "start": "node server.js"
+  },
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^1.7.4",
+    "blob-stream": "^0.1.3",
+    "body-parser": "^2.0.0",
+    "dotenv": "^17.0.0",
+    "express": "^5.0.0",
+    "js-yaml": "^4.1.0",
+    "json-to-tf": "^0.3.0",
+    "lazy-z": "1.12.1",
+    "next": "^15.5.4",
+    "regex-but-with-words": "^1.5.1",
+    "swagger-ui-express": "^5.0.0",
+    "tar-stream": "^3.1.6"
+  }
+}
