+++ unit-tests/api/schematics-api.test.js
@@ -1,6 +1,6 @@
 const { assert } = require("chai");
 const sinon = require("sinon");
-const controller = require("../../express-controllers/controller");
+const controller = require("../../server/express-controllers/controller");
 const res = require("../mocks/response.mock");
 const { initMockAxios } = require("lazy-z");
 const { state } = require("../../client/src/lib/state");
